refactor(rust): focus GCP provider modules
