+++ crates/source-runtime-next/src/gcp/iam/service_account_key/mod.rs
@@ -1,33 +1,19 @@
-//! Service-account-key wire decoding and Go-compatible normalization.
+//! Go-compatible family normalization.
+
+mod wire;
 
 use std::collections::BTreeMap;
 
-use serde::Deserialize;
 use serde_json::{Value, json};
 use time::OffsetDateTime;
 
 use super::{
-    GcpIamError, GcpIamFamily, GcpIamRecord, first_nonblank_gcp, insert_gcp_field,
-    normalized_gcp_time, normalized_observed_at, sanitize_gcp_event_id,
+    GcpIamError, GcpIamFamily, GcpIamRecord,
+    normalize::{
+        first_nonblank_gcp, insert_gcp_field, normalized_gcp_time, normalized_observed_at,
+        sanitize_gcp_event_id,
+    },
 };
-
-#[derive(Default, Deserialize)]
-#[serde(default, rename_all = "camelCase")]
-struct ServiceAccountKeyWire {
-    name: String,
-    #[serde(rename = "privateKeyType")]
-    _private_key_type: String,
-    #[serde(rename = "keyAlgorithm")]
-    _key_algorithm: String,
-    valid_after_time: String,
-    #[serde(rename = "validBeforeTime")]
-    _valid_before_time: String,
-    #[serde(rename = "keyOrigin")]
-    _key_origin: String,
-    #[serde(rename = "keyType")]
-    _key_type: String,
-    disabled: bool,
-}
 
 pub(super) fn normalize(
     payload: Value,
@@ -36,7 +22,7 @@ pub(super) fn normalize(
     service_account_email: &str,
     observed_at: OffsetDateTime,
 ) -> Result<GcpIamRecord, GcpIamError> {
-    let record: ServiceAccountKeyWire =
+    let record: wire::ServiceAccountKeyWire =
         serde_json::from_value(payload.clone()).map_err(|_| GcpIamError::InvalidResponse)?;
     let name = record.name.trim().to_owned();
     let provider_id = first_nonblank_gcp([name.as_str(), service_account_email])?;
